GEN_AI_Tasks/Task 3/code/requirements.txt
google-genai==1.66.0
beautifulsoup4==4.14.2
